refactor,feat(yt,ig,forms) Forms and form elements to own dirs. Updated source UI
New UI to switch between the sources (yt & ig) is better. Don't really
know how to describe it to be honest. I think its clever though, it
tries to convey the curent content source and is heavily animated so
it looks modern.
Refactored the sub froms and form element into their own dirs to
de-clutter the main project direcotry